Style: 스타일 분리. Text, Spacing 컴포넌트화
