+++ src/components/alertModal/index.js
@@ -1,0 +1,1 @@
+export { default as AlertModal } from './AlertModal';
