+++ src/components/mainModal/index.js
@@ -1,0 +1,1 @@
+export { default as MainModal } from './MainModal';
