+++ src/components/spacing/Spacing.style.jsx
@@ -1,0 +1,5 @@
+import styled from "styled-components";
+
+export const StyledSpacing = styled.div`
+  margin: ${(props) => props.margin || props.theme.spacing.md};
+`;
