+++ src/components/spacing/index.js
@@ -1,0 +1,1 @@
+export { default as Spacing } from './Spacing';
