+++ src/components/text/index.js
@@ -1,0 +1,1 @@
+export { default as Text } from './Text';
